Move CurvePoint to the new k256 module
